Commiting Audio Changes - Jay
Basically this is how Audio will work in the game
There will be one Audio GameObject and under it will be mutliple AudioManagers.
AudioManager scripts inherit the AbstractAudioManager script.
Currently There is the DollAudioManager and MusicManager.
All AudioManagers should be singleton's (you need to implement the singleton in each manager, couldnt do it in abstract duet to type casting problems) so you can call their methods from anywhere in the game.
AudioManager has some helpful helper functions like playOneShot and playLoopSound.

The MusicManager is a more special AudioManager that has some helpful functions for music control. Specifically, there is fadein and fadeout functions so we can start/stop music smoothly. I also implemented a switchsongs function so you can easily set to switch from song a to song b with a fade in fade out effect. To test this you can add the testSongSwitch script to one of the activate gameobjects.

There is also a new Audio folder. The files in here are divided into AudioSources and AudioClips. AudioClips are the raw sound files.

These are the specific things I have implemented
- Audio for walking
- Audio for ghost switching
- Music for main scene
- Temporary  music for  mechanical scenes (havent actually added to all the scenes yet).

Ask me for any questions :)