Fixed audio issues in the scene.
Added a new audio prefab as I realized people were adding audio wrong to the scene.
Now all you have to do is add the audio prefab to the scene and then add the appropriate loader prefab.

Added a mechanicalLevelLoader script and prefab for the mechanical levels. This plays the mechanical level theme from the start.