Fixed conveyor belt bug in 5a
fixed conveyor belt bug in 5a by adding a new script to dolls which
checks whether it has been moved in a direction by a conveyor belt yet
that frame, if it has then it will not let another conveyor belt move it
in said direction until the next frame. This is done by having a
constantly updating method setting variables saying whether a doll has
been moved by a conveyor belt false each update. When a conveyor belt
tries to move a doll, it only moves it if the variable is false, then
sets it to true after it moves the doll so no other belt can move it
until next frame.