Fixed Final scene UI: - Made all Final scene UI components scale to screen size. - Changed How we show text with a black/white background. Made it so it displays on a different panel (this way we are no scaling the panel, as this decreases the quality). - Added a fade in effect to "The End" text as well as a fade out effect. - Changed the font for the final scene as well as the other scenes (change the UI Canvas prefab)
NOTE: For some reason, I can't get the "Best Fit" feature for the text (where the font gets smaller when there is a lot of text) to work. I have looked into it a bit, and I can't find any real help online. I think it might be due to how we load the text
cause the font seems to scale when Im typing text into the scene.

Other changes:
- Fixed the final scene floor.
- Shortened a line in the final scene dialog to fit the textbox.
- Change the Conveyor belt speed as well as texture (did not use sprite-lit as this is really glitchy with the conveyor belt (basically two conveyor belt parts overlap and the light looks weird).