Fixed this bug about things still being selectable in the start screeen https://trello.com/c/dkCZh3sF/189-mouse-can-still-click-things-in-title-screen.
Also fixed a bug where in the start screen and in the title screen the selected objected would be unselected when minimizing or exiting the game and then coming back to it.
Basically we save and remember what the last selected object is, and re-select it when the selection becomes null.